requirements.txt
# static analysis
black

# code formatter
flake8

# static type checker
mypy

# others
Flask
SQLAlchemy
alembic
python-dotenv
requests
beautifulsoup4
schedule